Allow callers to perform the G'MIC image conversion
This allows more image layouts to be supported.
For example, the ImageSharp library expects callers to access the image
data one row at a time because it may not be able to store the entire
image in a managed array.

Added a `GrayAlpha88` value to the `NativeImageFormat` enumeration.
Removed the `Bgr888`, `Bgr888x`, `Bgra8888` and `Rgb888x` values from
the `NativeImageFormat` enumeration.